feat: medical record PUT requests:
 - com.safetynet.safetynetalertsapi.repositories.MedicalRecordRepository.java

 Format logger exception message;

Create update(MedicalRecord medicalRecord) method that:
 - retrieves the list of medical records from the data source file,
 - throws a ResourceNotFoundException if needed,
 - throws a NoCChangesDetectedException if the provided medical record is identical to the existing record,
 - calls the update() method of the JsonDataHandler.
 - returns the MedicalRecord to the MedicalRecordPersister.

 - com.safetynet.safetynetalertsapi.controllers.MedicalRecordController.java
Add updateMedicalRecord() method with PutMapping annotation, that accepts a MedicalRecordDto as request body and a PathVariable String identity.
This method calls the MedicalRecordPersister's updateMedicalRecord() method and returns a response entity if the process succeed, an 404 response if no matching medical record is found, or a 204 (no content) response if a NoChangesDetectedException is thrown.

com.safetynet.safetynetalertsapi.services.persisters.MedicalRecordPersister.java
Add updateMedicalRecord() method that maps the provided DTO to a domain object, calls the repository's update() method, and return the saved record or propagates an ResourceNotFoundException or a NoChangesDetectedException if needed.

JsonDataHandler.java
Handle medical records in the update() method.

DataSet.java
Fix typo (setMedicalrecords to setMedicalRecords).

NoChangesDetectedException.java
Add custom exception.